[refactor] rapid-grants page uses canonical marketing layout
Same alignment treatment as career-transition-grant: MarketingHero +
Breadcrumbs + Section-based body + centred Apply CTA. Stats and the
'See funded projects' scroll link move into a slim strip below the
breadcrumbs. Hardcoded hex (#356DB1, #2A5FA8, #7B8EA9, #52637B,
#CFE0F6, #D8E4F3) replaced with bluedot-* tokens.
